big refactoring, components are still not working. maybe move the mapper into the helpers?

+++ magnolia-handlebars/src/main/java/com/magnoliales/handlebars/rendering/renderer/HandlebarsRenderer.java
@@ -1,4 +1,4 @@
-package com.magnoliales.handlebars.templating.renderer;
+package com.magnoliales.handlebars.rendering.renderer;
 
 import com.github.jknack.handlebars.Context;
 import com.github.jknack.handlebars.Handlebars;
@@ -12,12 +12,10 @@ import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
 import com.github.jknack.handlebars.io.CompositeTemplateLoader;
 import com.github.jknack.handlebars.io.FileTemplateLoader;
 import com.github.jknack.handlebars.io.TemplateLoader;
-import com.magnoliales.handlebars.annotations.Component;
 import com.magnoliales.handlebars.mapper.NodeObjectMapper;
-import com.magnoliales.handlebars.setup.HandlebarsRegistry;
+import info.magnolia.cms.core.AggregationState;
 import info.magnolia.context.MgnlContext;
 import info.magnolia.jcr.node2bean.Node2BeanException;
-import info.magnolia.jcr.util.ContentMap;
 import info.magnolia.rendering.context.RenderingContext;
 import info.magnolia.rendering.engine.RenderException;
 import info.magnolia.rendering.engine.RenderingEngine;
@@ -34,22 +32,19 @@ import javax.jcr.Node;
 import javax.jcr.RepositoryException;
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class HandlebarsRenderer extends AbstractRenderer {
 
+    public static final String CURRENT_NODE_PROPERTY = "handlebars:node";
     private static final Logger logger = LoggerFactory.getLogger(HandlebarsRenderer.class);
 
     private Handlebars handlebars;
     private NodeObjectMapper nodeObjectMapper;
-    private HandlebarsRegistry handlebarsRegistry;
 
     @Inject
-    public HandlebarsRenderer(RenderingEngine renderingEngine, NodeObjectMapper nodeObjectMapper,
-                              HandlebarsRegistry handlebarsRegistry) {
+    public HandlebarsRenderer(RenderingEngine renderingEngine, NodeObjectMapper nodeObjectMapper) {
         super(renderingEngine);
         File templateDirectory = new File("src/main/resources/templates");
         TemplateLoader loader;
@@ -78,7 +73,6 @@ public class HandlebarsRenderer extends AbstractRenderer {
         }
 
         this.nodeObjectMapper = nodeObjectMapper;
-        this.handlebarsRegistry = handlebarsRegistry;
     }
 
     @Override
@@ -89,37 +83,21 @@ public class HandlebarsRenderer extends AbstractRenderer {
     @Override
     protected void onRender(Node node, RenderableDefinition renderableDefinition, RenderingContext renderingContext,
                             Map<String, Object> context, String templateScript) throws RenderException {
-
-        if (context.containsKey("components")) {
-            @SuppressWarnings("unchecked")
-            List<ContentMap> contentMaps = (List<ContentMap>) context.get("components");
-            for (ContentMap contentMap : contentMaps) {
-                try {
-                    Class<?> componentClass = Class.forName((String) contentMap.get("mgnl:template"));
-                    Node componentNode = contentMap.getJCRNode();
-                    Component component = componentClass.getAnnotation(Component.class);
-                    onRender(componentNode, renderableDefinition, renderingContext,
-                            new HashMap<String, Object>(), component.templateScript());
-                } catch (ClassNotFoundException e) {
-                    logger.error("Cannot render templateScript", e);
-                }
-            }
-        }
-        final AppendableWriter out;
+        Context combinedContext = null;
         try {
-            out = renderingContext.getAppendable();
-            Context combinedContext = Context.newBuilder(nodeObjectMapper.map(node))
-                    .combine("mgnl:node", node)
+            logger.info("Rendering node {} with {}", node.getPath(), templateScript);
+            combinedContext = Context.newBuilder(nodeObjectMapper.map(node))
+                    .combine(CURRENT_NODE_PROPERTY, node)
                     .resolver(JavaBeanValueResolver.INSTANCE, FieldValueResolver.INSTANCE, MapValueResolver.INSTANCE)
                     .build();
-            try {
-                Template template = handlebars.compile(templateScript);
-                template.apply(combinedContext, out);
-            } finally {
+            Template template = handlebars.compile(templateScript);
+            template.apply(combinedContext, renderingContext.getAppendable());
+        } catch (IOException | RepositoryException e) {
+            throw new RenderException(e);
+        } finally {
+            if (combinedContext != null) {
                 combinedContext.destroy();
             }
-        } catch (IOException e) {
-            logger.error("Cannot render templateScript", e);
         }
     }
 }
